Python/src/requirements.txt
semantic-kernel==1.35.0
python-dotenv==1.1.1
azure-search-documents==11.5.3
fastapi==0.116.1
pandas==2.3.1
uvicorn==0.35.0
streamlit>=1.47.0